Data Collection Manager | Test Architecture (#77)
* rough, incomplete implementation for DataCollectionManager

* remove core and merge missing files

* Delete duplicate package-lock.json

* fix logical deadlocks, setup error handling loop

* Started error checking

TODO: Create library for all errors we expect
TODO: add name to device adapters so that can be sent back

* Exceptions Library

First run with just 3 common used errors.  All they do is create AssertionErrors with a specific string. Also for identification with Error Handler

* Began utilizing ChariotExceptions Library

Added the few exceptions from ChariotExceptions to the Adapters and DataCollectionManager.
Also renamed ChariotException.py -> ChariotExceptions.py

* Started handler

Need a plan as to what to actually do. It might require some changes to the adapters as it needs an identifier, unless we can get the thread to do that.

* added deviceID to _generateStackTrace

Thread stack traces now include the device ID so error handler can work with that thread in particular

* quick error lib housekeeping

* fix conflict

* fix folder structure

* DeviceNotConnected error handled

Attempts to reconnect 3 times, and then stops collection for device if it can't reconnect.

* fixed how Chariot Exceptions were named

* cleaned up some missed Errors

* Added 'FailedToBeingCollectionError'

This error occurs when a device is unable to begin collecting data. The system will attempt to start again until giving up.

* Test Device started

Just started by copying what is needed and putting some stuff in.

* Test Device frame complete

Finished the frame for the Test Devices. Now needs to be tested and debuggeds, and have errors raise for ErrorHandler testing. Sample configs are also needed.

* refactor adapters to use private wrapped methods

the abstract DeviceAdapter will take care of generic work like managing if the device is connected or in a collection episode. This removes some code duplication in empirical device adapters.

* fix JSONTypes imports

* implement WorkerThread for consumers/producers to correctly handle errors

* pre-allocate buffer before filling in TestAdapter

* fix various type and logic errors

* add init files and mypy config to enable mypy checking via modules

* remove duplicate DeviceAdapter file

* type definition fixes

* cleanup on TestAdapter

* TestAdapter fixed local merge conflict

* stop passing errorQueue to devices: WorkerThread handles that

Also change OrderedDict to Dict as the order is not necessary.

* changes in prep of testing

* cleaned up more merge conflicts

* Fixed validation logic

* TestAdapter/TestConfiguration Unit tested

* Removed unused json module

* Changed getDeviceName to getId

* imported DatabaseWriter

* Updated test device to match new style

* WorkerThread refactor, fixes to DataCollectionManager based on testing

* Give DataCollectionManager repsonsibility in connecting and disconnecting from its DatabaseWriter

* update imports, add missing lib to requirements.txt

* connect and disconnect from devices in DataCollectionManager

* merge custom exceptions into one lib

* Updated imported modules to match new style

* updated imported modules to match new style

* Error handler can successfully reconnect a thread

* errorHandler started, but test fields are left in Device Adapter

* fix import, typedef

* DataCollectionManager refactor

* method name change: 'begin' to 'start'

* DataCollector refactor, error fixing and other improvements

* Initialize testing framework, write tests for chariot.collection module

* reduce cpu usage with poll delay calculations

* don't error out on repeated calls to stopDataCollection

Erroring out would be problematic when there is a race condition between a timer and a manual call to stop.

* update DataCollector tests to stress test

* implement modification locks during data collection

* adjust sleep for efficiency, fix databse write lock

* potential fix to test server/database issues

* definitive fix to server start race condition

* add test coverage

* let output thread sleep for max possible time

* Update DataCollector.py

* cleanup tests

* remove unused imports

* sleep for as long as possible when waiting for data from a device

* Update DataCollectionManager.py

* make database write lock private

* Type[List] cannot be typechecked at runtime

* parametrized types cannot be checked at runtime

* Update NetworkManager.py

Co-authored-by: kac486 <kac486@drexel.edu>