add devices to workers during a DCE
Added mechanisms to add devices to a Data Collection Worker during a DCE, and then allowing it to have data collected by a consumer.
This is neccessary as any time a device errors out, its thread dies and a new one is needed. The other alternative would be to make a whole new DataCollectionWorker process.
This also required the addition of  an identifier to data collection workers for use by the Error Handler