Producer Disconnect Error Handled
Undid a lot of the identifying changes made in the last commit. Could be ignored due to an earlier statement in  ErrorHandler method.

Handles disconnects by creating a new  device thread in the worker, this will attempt to reconnect. If it fails 3 times informs the user and then ignores that device. Uses a dictionary to track all devices.